+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 - Add `UluImageSlider`
   - This components uses UluSlider and replace functionality UluImageSlideShow had
 - Add `UluCaptionedFigure` to match frontend component
+- Add `UluAction` to be the unified API for how router-link/button/a are resolved (eg. to/href/click)
+  - This replaces repetitive duplicated code that resolves this throughout (UluButton, UluMenu, ...)
 
 ## 0.2.0-beta.16